release_2.5.6

+++ tests/test_v2_bucket.py
@@ -142,7 +142,7 @@ class TestBucket(TosTestBase):
     def test_bucket_with_acl(self):
         bucket_name = self.bucket_name + "-acl"
         for acl in ACLType:
-            if acl is not ACLType.ACL_Bucket_Owner_Entrusted:
+            if acl is not ACLType.ACL_Bucket_Owner_Entrusted and acl is not ACLType.ACL_Unknown:
                 self.client.create_bucket(bucket_name + acl.value, acl=acl)
                 self.client.delete_bucket(bucket_name + acl.value)
 
@@ -198,11 +198,11 @@ class TestBucket(TosTestBase):
         out = self.client.head_bucket(bucket=bucket_name)
         self.assertEqual(out.storage_class, StorageClassType.Storage_Class_Ia)
 
-        self.client.put_bucket_storage_class(bucket_name, StorageClassType.Storage_Class_Archive_Fr)
-        self.assertEqual(self.client.head_bucket(bucket_name).storage_class, StorageClassType.Storage_Class_Archive_Fr)
-
         self.client.put_bucket_storage_class(bucket_name, StorageClassType.Storage_Class_Standard)
         self.assertEqual(self.client.head_bucket(bucket_name).storage_class, StorageClassType.Storage_Class_Standard)
+
+        # self.client.put_bucket_storage_class(bucket_name, StorageClassType.Storage_Class_Archive_Fr)
+        # self.assertEqual(self.client.head_bucket(bucket_name).storage_class, StorageClassType.Storage_Class_Archive_Fr)
 
     def test_get_location(self):
         bucket_name = self.bucket_name + 'location'
